+++ tests/TestEnvironments/Php81/ValidatingHydratorTest.php
@@ -5,8 +5,15 @@ declare(strict_types=1);
 namespace Yiisoft\Hydrator\Validator\Tests\TestEnvironments\Php81;
 
 use PHPUnit\Framework\TestCase;
+use Yiisoft\Hydrator\AttributeHandling\ResolverFactory\ContainerAttributeResolverFactory;
+use Yiisoft\Hydrator\Hydrator;
+use Yiisoft\Hydrator\ObjectFactory\ContainerObjectFactory;
+use Yiisoft\Hydrator\Validator\Attribute\ValidateResolver;
 use Yiisoft\Hydrator\Validator\Tests\Support\TestHelper;
 use Yiisoft\Hydrator\Validator\Tests\TestEnvironments\Php81\Support\ValidateInput;
+use Yiisoft\Hydrator\Validator\ValidatingHydrator;
+use Yiisoft\Test\Support\Container\SimpleContainer;
+use Yiisoft\Validator\Validator;
 
 final class ValidatingHydratorTest extends TestCase
 {
@@ -42,5 +49,30 @@ final class ValidatingHydratorTest extends TestCase
         );
         $this->assertSame('y', $object->b);
         $this->assertSame('z', $object->c);
+    }
+
+    public function testValidateResolverCleanupAfterCreate1(): void
+    {
+        $validator = new Validator();
+        $validateResolver = new ValidateResolver($validator);
+        $container = new SimpleContainer(
+            [
+                ValidateResolver::class => $validateResolver,
+            ],
+            static fn(string $class) => new $class(),
+        );
+
+        $hydrator = new Hydrator(
+            attributeResolverFactory: new ContainerAttributeResolverFactory($container),
+        );
+        $validatingHydrator = new ValidatingHydrator($hydrator, $validator, $validateResolver);
+
+        $object = new ValidateInput();
+        $validatingHydrator->hydrate($object, ['b' => 'y', 'c' => 'z']);
+        $result = $object->getValidationResult();
+        $this->assertCount(1, $result->getErrors());
+
+        $hydrator->hydrate(new ValidateInput(), ['b' => 'y', 'c' => 'z']);
+        $this->assertCount(1, $result->getErrors());
     }
 }
